Rename files: interpolant.* -> interpolantmc.*
