+++ cosa2.cpp
@@ -28,7 +28,7 @@
 #include "bmc_simplepath.h"
 #include "defaults.h"
 #include "frontends/btor2_encoder.h"
-#include "interpolant.h"
+#include "interpolantmc.h"
 #include "kinduction.h"
 #include "printers/btor2_witness_printer.h"
 #include "prop.h"
